# docs/design/api-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "autohunt API",
    "description": "autohunt 系统对外 API 契约（技术设计 v1.3 §3）。\n\n通用约定：JSON；错误统一信封 `{\"error\": {\"code\", \"message\", \"details?\"}}`；时间一律 RFC3339（UTC 存储、前端本地化展示）；列表分页 `?cursor=&limit=`（默认 50）。\n\n鉴权（§3.1）：Web UI 使用 HttpOnly Cookie session（UISession）；外部 Agent 使用 `Authorization: Bearer ah_live_<random>`（AgentBearer）。同一套路由双鉴权，中间件按凭证类型打标 caller ∈ {ui, agent} 供状态机来源裁决（BR-11）。\n\n错误码表：UNAUTHORIZED(401)、FORBIDDEN(403)、NOT_FOUND(404)、PERMIT_REQUIRED(403)/PERMIT_INVALID(403)（submit_token 缺失或无效）、STATE_CONFLICT(409)（状态机裁决拒绝）、VALIDATION_ERROR(422)。\n\n契约 v2 增补（技设 §3.7，M3–M5 写侧，全部仅 UI session，事件详情只读除外）：简历上传/版本管理（/resumes，FR-1/2/3）、档案写（PUT /profile，FR-2）、邮箱账户绑定/解绑/状态/重授权（/email-accounts，FR-40/44）、事件确认/丢弃/修正与原文回溯（/events/{id}/confirm|discard|raw，FR-42/43，BR-2）、通知列表（/notifications，FR-32）、统计与 CSV 导出（/stats/*，FR-50/51/52，口径 §10.4）。\n\n契约 v2 修订（0.2.1）：确认流补 4 项——GET /confirmations 列表（仅 UI）、GET /confirmations/{id} 待确认变体按 caller 区分（UI 返回快照，Agent 仍仅 status）、POST /confirmations/{id}/close 手动关闭（仅 UI）、已确认响应含提交结果回写（submit_result/fail_reason/submitted_at，FR-24）；D-05 读侧三端点（/applications/{id}/history|confirmations|emails，FR-31/24/43）；提醒偏好 GET/PUT /settings/reminders（仅 UI）；GET /applications 增可选 from/to 筛选。既有 v1 冻结 19 端点语义未变动。",
    "version": "0.2.1"
  },
  "paths": {
...
    "/api/v1/applications/{application_id}/confirmations": {
      "get": {
        "tags": [
          "applications"
        ],
        "summary": "确认记录（FR-22/24，D-05 确认记录 Tab）【契约 v2 修订】",
        "description": "返回该投递关联的全部确认单（状态、确认时间、提交结果回写），按创建时间正序。",
        "operationId": "get_application_confirmations_api_v1_applications__application_id__confirmations_get",
        "parameters": [
          {
            "name": "application_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Application Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfirmationRecordList"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHORIZED — 未携带有效凭证",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "NOT_FOUND",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "UISession": []
          },
          {
            "AgentBearer": []
          }
        ]
      }
    },
...
    "/api/v1/confirmations": {
      "post": {
        "tags": [
          "confirmations"
        ],
        "summary": "创建待确认（Agent）",
        "description": "提交待确认字段-值快照。**响应不携带任何可提交许可（BR-1）。**",
        "operationId": "create_confirmation_api_v1_confirmations_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmationCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfirmationCreated"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHORIZED — 未携带有效凭证",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "200": {
            "description": "request_id 幂等命中：同一 Agent 重试返回首个确认单（AC-3 异常重试路径）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfirmationCreated"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "UISession": []
          },
          {
            "AgentBearer": []
          }
        ]
      },
      "get": {
        "tags": [
          "confirmations"
        ],
        "summary": "确认单列表（仅 UI，D-01 待确认分组/红点数据源）【契约 v2 修订】",
        "description": "返回确认单摘要列表（不含字段快照，快照走 GET /{id} 的 UI 变体）。筛选：status；分页 ?cursor=&limit=（默认 50）。Agent Bearer 一律 403。",
        "operationId": "list_confirmations_api_v1_confirmations_get",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ConfirmationStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfirmationList"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHORIZED — 未携带有效凭证",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "FORBIDDEN — 本端点仅接受 UI session 凭证；Agent Bearer 调用一律 403（BR-1 最后一道门，AC-3 负例）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "UISession": []
          }
        ]
      }
    },
    "/api/v1/confirmations/{confirmation_id}": {
      "get": {
        "tags": [
          "confirmations"
        ],
        "summary": "查询确认结果（Agent 轮询）",
        "description": "按状态与 caller 返回不同载荷：待确认 → Agent 仅 {status}，UI session 返回字段-值快照（fields + context，D-06 对照表，契约 v2 修订）；已驳回/已超时关闭 → status + reason?，流程终止；已确认 → status + confirmed_fields + submit_token + expires_at + 提交结果回写（submit_result / fail_reason / submitted_at，FR-24 结果视图）（token 过期/已消耗时 submit_token 为 null，走 UI「重新放行」恢复）。**BR-1 落地：submit_token 只在已确认时出现。**",
        "operationId": "get_confirmation_api_v1_confirmations__confirmation_id__get",
        "parameters": [
          {
            "name": "confirmation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Confirmation Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/ConfirmationPending"
                    },
                    {
                      "$ref": "#/components/schemas/ConfirmationPendingUI"
                    },
                    {
                      "$ref": "#/components/schemas/ConfirmationConfirmed"
                    },
                    {
                      "$ref": "#/components/schemas/ConfirmationClosed"
                    }
                  ],
                  "title": "Response Get Confirmation Api V1 Confirmations  Confirmation Id  Get"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHORIZED — 未携带有效凭证",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "NOT_FOUND",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "UISession": []
          },
          {
            "AgentBearer": []
          }
        ]
      }
    },
    "/api/v1/confirmations/{confirmation_id}/confirm": {
      "post": {
        "tags": [
          "confirmations"
        ],
        "summary": "人工确认（仅 UI）",
        "description": "用户在确认界面核对快照、可修改任意值后确认。服务端记录 confirmed_fields（含修改后值）与 confirmed_at，并在确认瞬间签发 submit_token：一次性、绑定 confirmation_id + confirmed_fields 哈希、TTL 30 分钟。",
        "operationId": "confirm_api_v1_confirmations__confirmation_id__confirm_post",
        "parameters": [
          {
            "name": "confirmation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Confirmation Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmationConfirm"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfirmationConfirmed"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHORIZED — 未携带有效凭证",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "FORBIDDEN — 本端点仅接受 UI session 凭证；Agent Bearer 调用一律 403（BR-1 最后一道门，AC-3 负例）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "NOT_FOUND",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "UISession": []
          }
        ]
      }
    },
    "/api/v1/confirmations/{confirmation_id}/reject": {
      "post": {
        "tags": [
          "confirmations"
        ],
        "summary": "人工驳回（仅 UI）",
        "description": "驳回后流程终止，状态为「已驳回」。",
        "operationId": "reject_api_v1_confirmations__confirmation_id__reject_post",
        "parameters": [
          {
            "name": "confirmation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Confirmation Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmationReject"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/ConfirmationPending"
                    },
                    {
                      "$ref": "#/components/schemas/ConfirmationPendingUI"
                    },
                    {
                      "$ref": "#/components/schemas/ConfirmationConfirmed"
                    },
                    {
                      "$ref": "#/components/schemas/ConfirmationClosed"
                    }
                  ],
                  "title": "Response Reject Api V1 Confirmations  Confirmation Id  Reject Post"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHORIZED — 未携带有效凭证",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "FORBIDDEN — 本端点仅接受 UI session 凭证；Agent Bearer 调用一律 403（BR-1 最后一道门，AC-3 负例）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "NOT_FOUND",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "UISession": []
          }
        ]
      }
    },
    "/api/v1/confirmations/{confirmation_id}/reissue": {
      "post": {
        "tags": [
          "confirmations"
        ],
        "summary": "重新放行（仅 UI）——token 过期/消耗后的唯一恢复路径",
        "description": "确认单处于「已确认」但 submit_token 已过期或已消耗（含回写失败被消耗）时，重新签发 token：原 confirmed_fields 不变、重新绑定哈希、重置 30 分钟 TTL。**Agent 侧不设任何换发/续期接口**；已回写成功的确认单不可重新放行（§3.4 步骤 5）。",
        "operationId": "reissue_api_v1_confirmations__confirmation_id__reissue_post",
        "parameters": [
          {
            "name": "confirmation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Confirmation Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfirmationConfirmed"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHORIZED — 未携带有效凭证",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "FORBIDDEN — 本端点仅接受 UI session 凭证；Agent Bearer 调用一律 403（BR-1 最后一道门，AC-3 负例）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "NOT_FOUND",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "STATE_CONFLICT — 确认单非「已确认」态、token 仍有效，或已回写成功（不可重新放行）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "UISession": []
          }
        ]
      }
    },
    "/api/v1/confirmations/{confirmation_id}/close": {
      "post": {
        "tags": [
          "confirmations"
        ],
        "summary": "手动关闭确认任务（仅 UI）【契约 v2 修订】",
        "description": "挂起无自动超时（PRD §12）的手动出口：待确认 → 已超时关闭，流程终止，留档可查。已确认的确认单不可经本端点关闭（走重新放行或回写闭环）。",
        "operationId": "close_api_v1_confirmations__confirmation_id__close_post",
        "parameters": [
          {
            "name": "confirmation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Confirmation Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/ConfirmationClose"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/ConfirmationPending"
                    },
                    {
                      "$ref": "#/components/schemas/ConfirmationPendingUI"
                    },
                    {
                      "$ref": "#/components/schemas/ConfirmationConfirmed"
                    },
                    {
                      "$ref": "#/components/schemas/ConfirmationClosed"
                    }
                  ],
                  "title": "Response Close Api V1 Confirmations  Confirmation Id  Close Post"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHORIZED — 未携带有效凭证",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "FORBIDDEN — 本端点仅接受 UI session 凭证；Agent Bearer 调用一律 403（BR-1 最后一道门，AC-3 负例）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "NOT_FOUND",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "STATE_CONFLICT — 确认单非「待确认」态（已确认/已驳回/已关闭不可手动关闭）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "UISession": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ConfirmationRecordList": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ConfirmationRecord"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "ConfirmationRecordList"
      },
...
      "ConfirmationStatus": {
        "type": "string",
        "enum": [
          "待确认",
          "已确认",
          "已驳回",
          "已超时关闭"
        ],
        "title": "ConfirmationStatus",
        "description": "§10.3 确认流状态。"
      },
...
      "ErrorEnvelope": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorEnvelope",
        "description": "统一错误信封（§3 通用约定）。"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
    }
  },
...
}